设计模式

+++ src/main/java/com/xwy/kkb/strategypattern/milkshopdemo3/Normal.java
@@ -1,14 +1,16 @@
 package com.xwy.kkb.strategypattern.milkshopdemo3;
 
 /**
- * @Classname Normal
- * @Created by 寂然
+ * @param
+ * @author xwy
  * @Description 正常收费
+ * @date 14/9/2021 下午6:21
+ * @return
  */
 public class Normal extends Promition {
 
     @Override
-    public int countPrice(int price) {
+    public double countPrice(int price) {
 
         //不打折
         return price;
